% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuLTM.R
\name{BuLTM}
\alias{BuLTM}
\title{Bayesian unidentified Linear Transformation model}
\usage{
BuLTM(
  Y,
  X,
  delta,
  probseries,
  K,
  a = 1,
  eta = 0.01,
  zeta = 1,
  v = 1,
  chains = 4,
  iter = 1500,
  warmup = 500,
  thin = 1,
  seed = 1234,
  adapt_delta = 0.8,
  max_treedepth = 9
)
}
\arguments{
\item{Y}{Numeric Vector of output values.}

\item{X}{Numeric Matrix of input values.}

\item{probseries}{Sequence from 0 to 1 or the step value of the sequence from 0 to 1.}

\item{K}{Integer value of truncated DP.}

\item{a}{Numeric value for mass parameter of DP.}

\item{eta}{hyperparameter eta.}

\item{zeta}{hyperparameter zeta.}

\item{v}{hyperparameter v.}

\item{chains}{Argument passed to \code{rstan::sampling}.}

\item{iter}{Argument passed to \code{rstan::sampling} .}

\item{warmup}{Argument passed to \code{rstan::sampling} .}

\item{thin}{Argument passed to \code{rstan::sampling} .}

\item{seed}{The seed for sampling.}

\item{adapt_delta}{The adapt_delta in Stan.}

\item{max_treedepth}{The max_treedepth in Stan.}
}
\value{
An object of class \code{BuLTM} returned by \code{rstan::sampling}
}
\description{
Bayesian unidentified Linear Transformation model
}
