% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuLTMfit.R
\name{BuLTMfit}
\alias{BuLTMfit}
\title{Unified nonparametric Bayesian inference model with Stan}
\usage{
BuLTMfit(
  Yi,
  X,
  delta,
  probseries,
  K,
  alpha,
  prior = "Exp",
  eta = 1,
  psi = 1,
  v = 1,
  chains = 4,
  iter = 2500,
  warmup = 500,
  thin = 1
)
}
\arguments{
\item{Yi}{Numeric Vector of output values}

\item{X}{Numeric Matrix of input values}

\item{probseries}{Sequence from 0 to 1 or the step value of the sequence from 0 to 1}

\item{K}{Integer value of truncated DP}

\item{alpha}{Numeric value for mass parameter of DP}

\item{prior}{The prior for coefficient, defaut is 'Exp', exponential; otherwise truncated Gaussian}

\item{chains}{Argument passed to \code{rstan::sampling}.}

\item{iter}{Argument passed to \code{rstan::sampling} .}

\item{warmup}{Argument passed to \code{rstan::sampling} .}

\item{thin}{Argument passed to \code{rstan::sampling} .}
}
\value{
An object of class \code{BuLTMfit} returned by \code{rstan::sampling}
}
\description{
Unified nonparametric Bayesian inference model with Stan
}
