% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuLTMfit2.R
\name{BuLTMfit2}
\alias{BuLTMfit2}
\title{Unified nonparametric Bayesian inference model with Stan}
\usage{
BuLTMfit2(
  Yi,
  X,
  delta,
  num_knots = 15,
  K,
  alpha,
  chains = 4,
  iter = 2500,
  warmup = 500,
  thin = 1
)
}
\arguments{
\item{Yi}{Numeric Vector of output values}

\item{X}{Numeric Matrix of input values}

\item{num_knots}{number of knots}

\item{K}{Integer value of truncated DP}

\item{alpha}{Numeric value for mass parameter of DP}

\item{chains}{Argument passed to \code{rstan::sampling}.}

\item{iter}{Argument passed to \code{rstan::sampling} .}

\item{warmup}{Argument passed to \code{rstan::sampling} .}

\item{thin}{Argument passed to \code{rstan::sampling} .}
}
\value{
An object of class \code{BuLTMfit2} returned by \code{rstan::sampling}
}
\description{
Unified nonparametric Bayesian inference model with Stan
}
