% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BuLTMfit2.R
\name{predict.BuLTMfit2}
\alias{predict.BuLTMfit2}
\title{Predict conditional survival functions for Unified nonparametric Bayesian inference model}
\usage{
\method{predict}{BuLTMfit2}(object, X = NULL, grids = NULL)
}
\arguments{
\item{object}{The BuLTMfit object}

\item{X}{vector of covariates}

\item{grids}{The time grids to be predicted}
}
\value{
A list including the estimated curve and 95\% intervals
}
\description{
Predict conditional survival functions for Unified nonparametric Bayesian inference model
}
