% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuffInform.R
\name{SuffitInf}
\alias{SuffitInf}
\title{Functions for sufficient informativeness criterion}
\usage{
SuffitInf(object, plot = TRUE)
}
\arguments{
\item{object}{The BuLTM object}

\item{plot}{The logical variable for plot}
}
\description{
Functions for sufficient informativeness criterion
}
